requirements.txt
absl-py==1.4.0
anyio==4.0.0
apify-client==1.4.1
apify-shared==1.0.2
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
